Stop searching when the optimal program is found
Since programs are generated in ascending order of length, the first
program to be found that produces the right state will be the optimal
one. Thus there's no reason to keep searching once a program has been
found.

This improves performance quite a bit in some cases and removes the
need to guess the optimal length to provide a short maximum length.

Additionally, this removes the program cache as it would only help
when reusing results from previous searches, which is currently being
done and wouldn't be correct when using different options (like
max_value) anyway.