Rework how programs are represented, parsed, run and generated
The program representation now consists of classes with a `__str__`
method, so `assembler.output` is no longer needed.

More importantly, a program now contains the information how much
memory it needs, removing the need to manually set this parameter and
removing the possibility of accidentally using a wrong value.

The opcodes are now represented as strings rather than functions from
the CPU, so that the program representation is no longer tied to the
CPU. This removes the need to instantiate dummy CPU objects when you
don't actually want to run a program and more importantly makes this
representation more usable when working with the program independently
of the CPU.

The program representation is now defined in the instruction_set
module, which removes the dependency from the assembler module to the
CPU module. This made it possible to move the `run` helper method into
the CPU module (where it belongs) without introducing a cyclical
dependency).

The CPU, assembler and program generator have been updated
according to these changes.

The assembler is now also a bit more robust. It now allows leading
white space and produces errors on invalid lines, including
non-existant opcodes, wrong numbers of arguments or invalid memory
addresses. The arity checks and the tracking of the memory size
happens based on the definitions in instruction_set and when new
instructions are added there, the assembler should be able to parse
and check those new instructions without needing to be modified.

Similarly the code generator now works solely from the defintions in
`instruction_set` and should likewise work without modification when
new instructions are added.